sorting static methods.
